Extract shared connection schema test helpers
Moves common collection and subtest logic from the per-connection-type
test files into a shared helpers file (connection_schema_test_helpers_test.go),
parameterised by blockKey and an entity-inclusion predicate. Both
elasticsearch_connection_schema_test.go and
kibana_connection_schema_test.go are updated to delegate to the shared
helpers, eliminating the duplicated collection loops and subtest
scaffolding.
